Only store parameter overrides on the installation record
Installation.Parameters should reflect parameter values that the user
selected, similar to --param on the command line. It shouldn't contain
any resolved values, such as from parameter sets as we already track
those on the Installation.ParameterSets field.
