+++ pkg/porter/cnab.go
@@ -239,6 +239,10 @@ func (o *bundleFileOptions) validateCNABFile(cxt *context.Context) error {
 // LoadParameters validates and resolves the parameters and sets. It must be
 // called after porter has loaded the bundle definition.
 func (o *sharedOptions) LoadParameters(p *Porter) error {
+	if o.combinedParameters != nil {
+		return nil
+	}
+
 	err := o.parseParams()
 	if err != nil {
 		return err
