+++ pkg/porter/lifecycle.go
@@ -154,26 +154,29 @@ func (p *Porter) BuildActionArgs(installation claims.Installation, action Bundle
 		}
 	}
 
+	// Resolve the final set of typed parameters, taking into account the user overrides, parameter sources
+	// and defaults
+	err = opts.LoadParameters(p)
+	if err != nil {
+		return cnabprovider.ActionArguments{}, err
+	}
+	resolvedParams, err := p.resolveParameters(installation, bundleRef.Definition, action.GetAction(), opts.combinedParameters)
+	if err != nil {
+		return cnabprovider.ActionArguments{}, err
+	}
+
 	args := cnabprovider.ActionArguments{
 		Action:                action.GetAction(),
 		Installation:          installation,
 		BundleReference:       bundleRef,
-		Params:                make(map[string]string, len(opts.combinedParameters)), // TODO(carolynvs): set on the installation record
+		Params:                resolvedParams,
 		CredentialIdentifiers: make([]string, len(opts.CredentialIdentifiers)),
 		Driver:                opts.Driver,
 		AllowDockerHostAccess: opts.AllowAccessToDockerHost,
 	}
 
-	err = opts.LoadParameters(p)
-	if err != nil {
-		return cnabprovider.ActionArguments{}, err
-	}
-
 	// Do a safe copy so that modifications to the args aren't also made to the
 	// original options, which is confusing to debug
-	for k, v := range opts.combinedParameters {
-		args.Params[k] = v
-	}
 	copy(args.CredentialIdentifiers, opts.CredentialIdentifiers)
 
 	return args, nil
