+++ pkg/porter/parameters.go
@@ -1,11 +1,13 @@
 package porter
 
 import (
+	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"sort"
 	"time"
 
+	"get.porter.sh/porter/pkg/claims"
 	"get.porter.sh/porter/pkg/cnab"
 	"get.porter.sh/porter/pkg/editor"
 	"get.porter.sh/porter/pkg/encoding"
@@ -14,6 +16,8 @@ import (
 	"get.porter.sh/porter/pkg/printer"
 	"get.porter.sh/porter/pkg/secrets"
 	"get.porter.sh/porter/pkg/storage"
+	"github.com/cnabio/cnab-go/bundle"
+	"github.com/cnabio/cnab-go/bundle/definition"
 	"go.mongodb.org/mongo-driver/bson"
 
 	dtprinter "github.com/carolynvs/datetime-printer"
@@ -492,4 +496,156 @@ func (p *Porter) ParametersApply(o ApplyOptions) error {
 	}
 
 	return p.Parameters.UpsertParameterSet(params)
+}
+
+// resolveParameters accepts a set of parameter assignments and combines them
+// with parameter sources and default parameter values to create a full set
+// of parameters.
+func (p *Porter) resolveParameters(installation claims.Installation, bun cnab.ExtendedBundle, action string, params map[string]string) (map[string]interface{}, error) {
+	mergedParams := make(secrets.Set, len(params))
+	paramSources, err := p.resolveParameterSources(bun, installation)
+	if err != nil {
+		return nil, err
+	}
+
+	for key, val := range paramSources {
+		mergedParams[key] = val
+	}
+
+	// Apply user supplied parameter overrides last
+	for key, rawValue := range params {
+		param, ok := bun.Parameters[key]
+		if !ok {
+			return nil, fmt.Errorf("parameter %s not defined in bundle", key)
+		}
+
+		def, ok := bun.Definitions[param.Definition]
+		if !ok {
+			return nil, fmt.Errorf("definition %s not defined in bundle", param.Definition)
+		}
+
+		// Apply porter specific conversions, like retrieving file contents
+		value, err := p.getUnconvertedValueFromRaw(bun, def, key, rawValue)
+		if err != nil {
+			return nil, err
+		}
+
+		mergedParams[key] = value
+	}
+
+	// Now convert all parameters which are currently strings into the
+	// proper type for the parameter, e.g. "false" -> false
+	typedParams := make(map[string]interface{}, len(mergedParams))
+	for key, unconverted := range mergedParams {
+		param, ok := bun.Parameters[key]
+		if !ok {
+			return nil, fmt.Errorf("parameter %s not defined in bundle", key)
+		}
+
+		def, ok := bun.Definitions[param.Definition]
+		if !ok {
+			return nil, fmt.Errorf("definition %s not defined in bundle", param.Definition)
+		}
+
+		if def.Type != nil {
+			value, err := def.ConvertValue(unconverted)
+			if err != nil {
+				return nil, errors.Wrapf(err, "unable to convert parameter's %s value %s to the destination parameter type %s", key, unconverted, def.Type)
+			}
+			typedParams[key] = value
+		} else {
+			// bundle dependency parameters can be any type, not sure we have a solid way to do a typed conversion
+			typedParams[key] = unconverted
+		}
+
+	}
+
+	return bundle.ValuesOrDefaults(typedParams, &bun.Bundle, action)
+}
+
+func (p *Porter) getUnconvertedValueFromRaw(b cnab.ExtendedBundle, def *definition.Schema, key, rawValue string) (string, error) {
+	// the parameter value (via rawValue) may represent a file on the local filesystem
+	if b.IsFileType(def) {
+		if _, err := p.FileSystem.Stat(rawValue); err == nil {
+			bytes, err := p.FileSystem.ReadFile(rawValue)
+			if err != nil {
+				return "", errors.Wrapf(err, "unable to read file parameter %s", key)
+			}
+			return base64.StdEncoding.EncodeToString(bytes), nil
+		}
+	}
+	return rawValue, nil
+}
+
+func (p *Porter) resolveParameterSources(bun cnab.ExtendedBundle, installation claims.Installation) (secrets.Set, error) {
+	if !bun.HasParameterSources() {
+		if p.Debug {
+			fmt.Fprintln(p.Err, "No parameter sources defined, skipping")
+		}
+		return nil, nil
+	}
+
+	if p.Debug {
+		fmt.Fprintln(p.Err, "Resolving parameter sources...")
+	}
+
+	parameterSources, err := bun.ReadParameterSources()
+	if err != nil {
+		return nil, err
+	}
+
+	values := secrets.Set{}
+	for parameterName, parameterSource := range parameterSources {
+		if p.Debug {
+			fmt.Fprintln(p.Err, "Resolving parameter source", parameterName)
+		}
+		for _, rawSource := range parameterSource.ListSourcesByPriority() {
+			var installationName string
+			var outputName string
+			switch source := rawSource.(type) {
+			case cnab.OutputParameterSource:
+				installationName = installation.Name
+				outputName = source.OutputName
+			case cnab.DependencyOutputParameterSource:
+				// TODO(carolynvs): does this need to take namespace into account
+				installationName = cnab.BuildPrerequisiteInstallationName(installation.Name, source.Dependency)
+				outputName = source.OutputName
+			}
+
+			output, err := p.Claims.GetLastOutput(installation.Namespace, installationName, outputName)
+			if err != nil {
+				// When we can't find the output, skip it and let the parameter be set another way
+				if errors.Is(err, storage.ErrNotFound{}) {
+					if p.Debug {
+						fmt.Fprintf(p.Err, "No previous output found for %s from %s/%s\n", outputName, installation.Namespace, installationName)
+					}
+					continue
+				}
+				// Otherwise, something else has happened, perhaps bad data or connectivity problems, we can't ignore it
+				return nil, errors.Wrapf(err, "could not set parameter %s from output %s of %s", parameterName, outputName, installation)
+			}
+
+			param, ok := bun.Parameters[parameterName]
+			if !ok {
+				return nil, fmt.Errorf("parameter %s not defined in bundle", parameterName)
+			}
+
+			def, ok := bun.Definitions[param.Definition]
+			if !ok {
+				return nil, fmt.Errorf("definition %s not defined in bundle", param.Definition)
+			}
+
+			if bun.IsFileType(def) {
+				values[parameterName] = base64.StdEncoding.EncodeToString(output.Value)
+			} else {
+				values[parameterName] = string(output.Value)
+			}
+
+			if p.Debug {
+				fmt.Fprintf(p.Out, "Injected installation %s output %s as parameter %s\n", installation, outputName, parameterName)
+			}
+		}
+	}
+
+	return values, nil
 }
